+++ cobras/server/app.py
@@ -212,9 +212,12 @@ class AppRunner():
         db.close()
         asyncio.ensure_future(db.wait_closed())
 
-    def terminate(self):
-        asyncio.get_event_loop().run_until_complete(self.cleanup())
-
+    async def closeServer(self):
         # Now close websocket server
         self.server.close()
-        self.server.wait_closed()
+        await self.server.wait_closed()
+
+    def terminate(self):
+        asyncio.get_event_loop().run_until_complete(self.cleanup())
+        asyncio.get_event_loop().run_until_complete(self.closeServer())
+